+++ seeders/districtSeeder.go
@@ -26,7 +26,7 @@ func SeedDistrict(db *gorm.DB) {
 	for i := 0; i < 100; i++ {
 		uProvID, _ := strconv.ParseUint(response.RajaOngkir.CityResults[i].ProvinceID, 10, 32)
 		dist.Name = response.RajaOngkir.CityResults[i].CityName
-		dist.ProvinceID = uProvID
+		dist.ProvinceID = uint(uProvID)
 		dist.ID = 0
 		db.Create(&dist)
 	}
